requirements.txt
langchain
python-dotenv
ipykernel 
langchain_community
pypdf
bs4
langchain-text-splitters
wikipedia
langchain-openai
chromadb
langchain_huggingface
faiss-cpu
sentence-transformers
langchain-groq
streamlit
langchain-ollama
mysql-connector-python
SQLAlchemy